third_party/WebKit/LayoutTests/printing/webgl-repeated-printing-expected.txt

 PASS successfullyParsed is true

TEST COMPLETE
Test 0: canvas should be green
PASS pixel[0] is within 1 of 0
PASS pixel[1] is within 1 of 255
PASS pixel[2] is within 1 of 0
Test 1: canvas should be red
PASS pixel[0] is within 1 of 255
PASS pixel[1] is within 1 of 0
PASS pixel[2] is within 1 of 0
Test 2: canvas should be blue
PASS pixel[0] is within 1 of 0
PASS pixel[1] is within 1 of 0
PASS pixel[2] is within 1 of 255
